Update `publishBlock` to disambiguate meaning of `success` in broadcast
The current description says:

> and a successful response (20X) only indicates that the broadcast has
been successful

"Successful" can be interpreted in two ways:

1. the block was sent to neighbouring peers (ie the node could parse the
`json`, more or less).
2. the block passed gossip validation and was then sent to neighbouring
peers.

If a beacon node gossips blocks that fail gossip validation, they stand
to lose peer score and may be disconnected from the network - as such,
publishing an invalid block poses a risk to all validators using the
same node.

This PR resolves the ambiguity by catering to the latter interpretation
and assumes gossip validation is performed on blocks before sending
them.

It also stresses that broadcast may be performed before integration so
as to aid timely delivery.