Clarify validator id as any bytes48
Validator ID in a query parameter should not be required to be on the g2 curve to be valid, any bytes48 should be allowed to be queried on.

Reading in any bytes48 without validating against the g2 curve will reduce overheads of parsing large lists, and given that valid missing keys are just not returned, there are no real negatives to querying a key that will definitely be missing.

fixes #183