Add 415 response as a reason to fail POSTs.
Now that block posts are able to accept SSZ input, it is useful to be able to fail them on the basis that the server doesn't support the nominated type.

A caveat is probably that JSON should be supported by all servers, but its possible that some servers don't yet support SSZ type.

The clinet is then able to determine that the server they're connected to doesn't support SSZ (for example) and stop attempting requests that can't succeed.