removed minItems, maxItems, added a comment in description of fields.
This does mean the limits won't be enforced in the openapi spec anymore, but the upside is that the example is now viewable, because it's only got 1 element in each array, so you can scroll easily, and it is fast to create.

There was also no need to use 'allOf', as there were no attributes being overridden, so cleaned up.