Made eth-consensus-version required on new blocks endpoints
When initially adding eth-consensus-version to interfaces it wasn't able to be required due to backwards compatibility issues.

Given these interfaces are new, we should set it as a required header. This will simplify the determination of whether we're reading a deneb blockContents or a pre-deneb signed block.

Potentially we deprecate the other interface once deneb comes around, but that's up for discussion, and I'm not putting that into this PR, just making this header required for the new interfaces.